Refactor engine hours handler
