+++ src/org/traccar/handler/EngineHoursHandler.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 Anton Tananaev (anton@traccar.org)
+ * Copyright 2018 - 2019 Anton Tananaev (anton@traccar.org)
  * Copyright 2018 Andrey Kunitsyn (andrey@traccar.org)
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -14,18 +14,26 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.traccar;
+package org.traccar.handler;
 
 import io.netty.channel.ChannelHandler;
+import org.traccar.BaseDataHandler;
+import org.traccar.database.IdentityManager;
 import org.traccar.model.Position;
 
 @ChannelHandler.Sharable
 public class EngineHoursHandler extends BaseDataHandler {
 
+    private final IdentityManager identityManager;
+
+    public EngineHoursHandler(IdentityManager identityManager) {
+        this.identityManager = identityManager;
+    }
+
     @Override
     protected Position handlePosition(Position position) {
         if (!position.getAttributes().containsKey(Position.KEY_HOURS)) {
-            Position last = Context.getIdentityManager().getLastPosition(position.getDeviceId());
+            Position last = identityManager.getLastPosition(position.getDeviceId());
             if (last != null) {
                 long hours = last.getLong(Position.KEY_HOURS);
                 if (last.getBoolean(Position.KEY_IGNITION) && position.getBoolean(Position.KEY_IGNITION)) {
